tweak(composition): delete scoreboard resource
This resource is outdated, known-insecure, problematic, and incompatible with modern standards. It was originally written in 2014 or so as a test, but never did much of note.

It should ideally be rewritten in a generic way, with an API to add columns from state bag entries or such, requesting players from the server (through a latent event) when opened, all the while being themeable similar to `chat` and documented in a minimal way as well.